Fix multi checkbox bug where specify doesnt remove required from others, fix unit scheduling bug, better and more Selenium tests
