+++ buildings/fixtures/evalunits.json
@@ -1,0 +1,92 @@
+[
+{
+    "model": "buildings.evalunit",
+    "pk": "id1",
+    "fields": {
+        "lat": 45.5208431059,
+        "lng": -73.5812809097,
+        "muni": "Montréal",
+        "muni_code": "66023",
+        "arrond": "REM21",
+        "address": "4288 - 4296 avenue Laval",
+        "num_adr_inf": "4288",
+        "num_adr_inf_2": null,
+        "num_adr_sup": "4296",
+        "num_adr_sup_2": null,
+        "street_name": "avenue Laval",
+        "apt_num": null,
+        "apt_num_1": null,
+        "apt_num_2": null,
+        "mat18": "984243502820000000",
+        "cubf": 1000,
+        "file_num": "30-F50054225",
+        "nghbr_unit": "2598",
+        "owner_date": "1987-12-09T05:00:00Z",
+        "owner_type": "physical",
+        "owner_status": "landowner",
+        "lot_lin_dim": 12.81,
+        "lot_area": 273.2,
+        "max_floors": 2,
+        "const_yr": null,
+        "const_yr_real": null,
+        "floor_area": null,
+        "phys_link": null,
+        "const_type": null,
+        "num_dwelling": 4,
+        "num_rental": null,
+        "num_non_res": null,
+        "apprais_date": "2018-07-01T04:00:00Z",
+        "lot_value": 495900,
+        "building_value": 353300,
+        "value": 849200,
+        "prev_value": 725800,
+        "associated": null,
+        "date_added": "2023-08-14T04:00:00Z"
+    }
+},
+{
+    "model": "buildings.evalunit",
+    "pk": "id2",
+    "fields": {
+        "lat": 45.520270497,
+        "lng": -73.5800229015,
+        "muni": "Montréal",
+        "muni_code": "66023",
+        "arrond": "REM21",
+        "address": "4210 avenue Laval",
+        "num_adr_inf": "4210",
+        "num_adr_inf_2": null,
+        "num_adr_sup": null,
+        "num_adr_sup_2": null,
+        "street_name": "avenue Laval",
+        "apt_num": null,
+        "apt_num_1": null,
+        "apt_num_2": null,
+        "mat18": "984252486420000000",
+        "cubf": 1000,
+        "file_num": "30-F50052650",
+        "nghbr_unit": "2598",
+        "owner_date": "2013-12-10T05:00:00Z",
+        "owner_type": "physical",
+        "owner_status": "landowner",
+        "lot_lin_dim": 7.87,
+        "lot_area": 167.9,
+        "max_floors": 2,
+        "const_yr": 1857,
+        "const_yr_real": "E",
+        "floor_area": 202.2,
+        "phys_link": "row house",
+        "const_type": "full-storey",
+        "num_dwelling": 4,
+        "num_rental": null,
+        "num_non_res": null,
+        "apprais_date": "2018-07-01T04:00:00Z",
+        "lot_value": 304700,
+        "building_value": 400300,
+        "value": 705000,
+        "prev_value": 602600,
+        "associated": null,
+        "date_added": "2023-08-14T04:00:00Z"
+    }
+}
+]
