+++ buildings/fixtures/user.json
@@ -1,0 +1,20 @@
+[
+{
+    "model": "buildings.user",
+    "pk": 4,
+    "fields": {
+        "password": "pbkdf2_sha256$390000$nz78rqRXmqXHL0K997qdyp$Le4KeZR3iLuvgV5eoj+PiD+peFpRZzgDQzSQLpoAJIo=",
+        "last_login": null,
+        "is_superuser": false,
+        "username": "testuser",
+        "first_name": "",
+        "last_name": "",
+        "email": "test@test.com",
+        "is_staff": false,
+        "is_active": true,
+        "date_joined": "2023-08-16T19:47:53.380Z",
+        "groups": [],
+        "user_permissions": []
+    }
+}
+]
